Arel::Nodes::TypeCast
