+++ lib/arel/extensions.rb
@@ -25,6 +25,7 @@ require 'arel/extensions/ordering'
 require 'arel/extensions/all'
 require 'arel/extensions/any'
 require 'arel/extensions/array_subselect'
+require 'arel/extensions/type_cast'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
